% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brainAtlas}
\alias{brainAtlas}
\title{Harvard-Oxford brain atlas}
\format{
A numeric matrix with 48 rows (cortical areas) and 48900 columns
  (voxels). Non-zero entries in row \eqn{r} indicate the voxels in the
  \eqn{r}-th region. The names of the areas are stored in the row names of
  the matrix.
}
\source{
\url{https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/Atlases}
}
\usage{
brainAtlas
}
\description{
Probabilistic brain atlases covering 48 cortical structural areas, derived
from structural data and segmentations.
}
\details{
Obtained via the Python module [nilearn](http://nilearn.github.io/),
  thanks to the R package [reticulate](https://rstudio.github.io/reticulate/)
}
\keyword{datasets}
