% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulkRNAseq}
\alias{bulkRNAseq}
\title{Bulk RNAseq matrix of the Cancer Genome Atlas Urothelial Bladder Carcinoma (TCGA-BLCA)}
\format{
A numeric matrix with :20146 rows (gens) and 135 columns
  (cancer samples). The samples are classified into two subgoups:
 \enumerate{
 \item{65 samples are in stage ii of an Urothelial Bladder Carcinoma cancer}
 \item{70 samples are in stage iii of an Urothelial Bladder Carcinoma cancer}
 }
 The column names provide the mutational status ("stage ii" or "stage iii").
 The row names are gene names
}
\source{
\url{https://www.cancer.org/cancer/bladder-cancer/detection-diagnosis-staging/staging.html}
}
\usage{
bulkRNAseq
}
\description{
Bulk RNAseq matrix of the Cancer Genome Atlas Urothelial Bladder Carcinoma (TCGA-BLCA)
}
\details{
Obtained via R packages [curatedTCGAData](https://bioconductor.org/packages/release/data/experiment/html/curatedTCGAData.html)
 and [TCGAutils](https://bioconductor.org/packages/release/bioc/html/TCGAutils.html)
}
\references{
Sabina Chiaretti, Xiaochun Li, Robert Gentleman, Antonella Vitale, Marco Vignetti, Franco Mandelli, Jerome Ritz, and Robin Foa Gene expression profile of adult T-cell acute lymphocytic leukemia identifies distinct subsets of patients with different response to therapy and survival. Blood, 1 April 2004, Vol. 103, No. 7.
}
\keyword{datasets}
