% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COPD_expr}
\alias{COPD_expr}
\title{Gene expression profiles of 135 samples of chronic obstructive pulmonary disease.}
\format{
A numeric matrix with 12,531 rows (genes) and 135 columns ( samples).
}
\source{
\url{https://bioconductor.org/packages/GEOquery/}
}
\usage{
COPD_expr
}
\description{
A dataset containing sequencing-based gene expression measurements from a
study on chronic obstructive pulmonary disease (COPD).
}
\details{
Obtained via the Bioconductor package GEOquery
}
\references{
Bahr, Timothy M, Grant J Hughes, Michael Armstrong, Rick Reisdorph, Christopher D Coldren, Michael G Edwards, Christina Schnell, et al. 2013. “Peripheral Blood Mononuclear Cell Gene Expression in Chronic Obstructive Pulmonary Disease.” American Journal of Respiratory Cell and Molecular Biology 49 (2): 316–23.
}
\keyword{datasets}
