% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expr_ALL_annotation}
\alias{expr_ALL_annotation}
\title{Probe annotation data for the hg_u95av2 Affymetrix microarray}
\format{
A numeric matrix with 10495 rows (probe sets) and 6 columns:
 \describe{
 \item{affy_hg_u95av2}{Affymetrix probe set id, e.g. "1000_at"}
 \item{hgnc_symbol}{HUGO gene symbol, e.g. "MAPK3"}
 \item{chromosome_name}{character value in "1":"22" or "X", "Y"}
 \item{start_position}{start position of the probe set (in bp)}
 \item{end_position}{start position of the probe set (in bp)}
 \item{band}{chromosomal band, e.g. "p11.2"}
 }
}
\source{
\url{https://bioconductor.org/packages/release/data/annotation/html/hgu95av2.html}
}
\usage{
expr_ALL_annotation
}
\description{
Probe annotation data for Affymetrix Human Genome U95 Set.
}
\keyword{datasets}
