% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expr_COPD_design}
\alias{expr_COPD_design}
\title{Design matrix  of COPD dataset.}
\format{
A numeric matrix with 135 rows (samples) and 11 columns (variables).
  The variables are:
 \describe{
 \item{group}{charactor. Two modalities variable coding "COPD" and "Control"}
 \item{age_enroll}{Numeric. Age of the participant (in years) at the time of enrollment in the study.}
  \item{ats_packyears}{Numeric. Smoking history expressed in pack-years, calculated according to ATS guidelines.}
  \item{bmi}{Numeric. Body mass index (BMI).}
  \item{distwalked}{Numeric. Distance walked (in meters) during the 6-minute walk test (6MWT),
  a standard measure of exercise capacity in COPD patients.}
  \item{fev1_fvc_utah}{Numeric. Ratio of forced expiratory volume in 1 second (FEV₁) to forced vital capacity (FVC),
  measured at the University of Utah. }
  \item{fev1pp_utah}{Numeric. FEV₁ expressed as a percentage of the predicted value, measured at the University of Utah.
  Used to assess disease severity.}
  \item{finalgold}{Integer. GOLD stage of COPD severity based on post-bronchodilator FEV₁}
  \item{gender}{Binary (0/1). Biological sex of the participant: 0 = female, 1 = male.}
  \item{parentalcopd}{Integer. Number of biological parents with a diagnosis of COPD}
  \item{smokcignow}{Binary (0/1). Current smoking status: 1 = current smoker, 0 = non-smoker or former smoker.}
 }
}
\source{
\url{https://bioconductor.org/packages/GEOquery/}
}
\usage{
expr_COPD_design
}
\description{
A dataset containing the design matrix from a study on chronic obstructive
pulmonary disease (COPD).
}
\details{
Obtained via the Bioconductor package GEOquery
}
\references{
Bahr, Timothy M, Grant J Hughes, Michael Armstrong, Rick
  Reisdorph, Christopher D Coldren, Michael G Edwards, Christina Schnell, et
  al. 2013. “Peripheral Blood Mononuclear Cell Gene Expression in Chronic
  Obstructive Pulmonary Disease.” American Journal of Respiratory Cell and
  Molecular Biology 49 (2): 316–23.
}
\seealso{
[expr_COPD]
}
\keyword{datasets}
