% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fMRI_localizer}
\alias{fMRI_localizer}
\title{fMRI data for 30 samples from the Localizer data set}
\format{
A numeric matrix with 48900 rows (voxels) and 30 columns (samples).
The samples are classified into two subgoups ("left click" vs "right click"), corresponding to two different motor tasks the subjects were asked to perform.
}
\source{
\url{http://brainomics.cea.fr/localizer/}
}
\usage{
fMRI_localizer
}
\description{
Functional Magnetic Resonance Imaging (fMRI) measurements from the Localizer data set
}
\details{
To import this data we have used the Python module [nilearn](http://nilearn.github.io/)
}
\references{
Orfanos, D. P., Michel, V., Schwartz, Y., Pinel, P., Moreno, A., Le Bihan, D., & Frouin, V. (2017). The brainomics/localizer database. Neuroimage, 144, 309-314.

  Abraham, A., Pedregosa, F., Eickenberg, M., Gervais, P., Mueller, A., Kossaifi, J., Gramfort, A., Thirion, B. and Varoquaux, G., 2014. Machine learning for neuroimaging with scikit-learn. Frontiers in neuroinformatics, 8, p.14.
}
\keyword{datasets}
