% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgu95av2_GO_BP}
\alias{hgu95av2_GO_BP}
\title{Subset of Gene Ontology terms for the hgu95av2 microarray}
\format{
A numeric matrix with 12625 rows (microarray probes) and 83 columns
  (GO terms from the Biological Process ontology). Non-zero entries in
  column \eqn{c} indicate the probes corresponding to the \eqn{c}-th GO
  term. The GO ids are stored in the column names of the matrix.
}
\source{
\url{https://bioconductor.org/packages/hgu95av2.db/}
}
\usage{
hgu95av2_GO_BP
}
\description{
Subset of Gene Ontology terms for the hgu95av2 microarray
}
\details{
Obtained via the Bioconductor packages topGO, hgu95av2.db
}
\keyword{datasets}
