% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq_blca_GO}
\alias{RNAseq_blca_GO}
\title{Selected Gene Ontology terms for the RNAseq_blca data set}
\format{
A numeric matrix with 20196 rows (genes) and 362 columns (selected GO
  terms). Non-zero entries in column \eqn{c} indicate the genes
  corresponding to the \eqn{c}-th GO term. The GO ids are stored in the
  column names of the matrix.
}
\source{
\url{https://bioconductor.org/packages/org.Hs.eg.db/}
}
\usage{
RNAseq_blca_GO
}
\description{
Selected Gene Ontology terms for the RNAseq_blca data set
}
\details{
Obtained via the Bioconductor package org.Hs.eg.db
}
\keyword{datasets}
