% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expr_ALL_annotation}
\alias{expr_ALL_annotation}
\title{Gene annotation data for the expr_ALL data set}
\format{
A numeric matrix with 9038 rows (genes) and 5 columns:
 \describe{
 \item{ensembl_gene_id}{ENSEMBL gene id, e.g. "ENSG00000000003"}
 \item{chromosome_name}{character value in "1":"22" or "X", "Y"}
 \item{start_position}{start position of the probe set (in bp)}
 \item{end_position}{start position of the probe set (in bp)}
 \item{band}{chromosomal band, e.g. "p11.2"}
 }
}
\source{
\url{https://bioconductor.org/packages/release/data/annotation/html/hgu95av2.html}
}
\usage{
expr_ALL_annotation
}
\description{
Gene annotation data for the expr_ALL data set
}
\keyword{datasets}
