% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAseq_blca}
\alias{RNAseq_blca}
\title{Bulk RNAseq matrix of the Cancer Genome Atlas Urothelial Bladder Carcinoma (TCGA-BLCA)}
\format{
A numeric matrix with :20146 rows (gens) and 135 columns
  (cancer samples). The samples are classified into two subgoups:
 \enumerate{
 \item{65 samples are in stage ii of an Urothelial Bladder Carcinoma cancer}
 \item{70 samples are in stage iii of an Urothelial Bladder Carcinoma cancer}
 }
 The column names provide the mutational status ("stage ii" or "stage iii").
 The row names are gene names
}
\source{
\url{https://www.cancer.org/cancer/bladder-cancer/detection-diagnosis-staging/staging.html}
}
\usage{
RNAseq_blca
}
\description{
Bulk RNAseq matrix of the Cancer Genome Atlas Urothelial Bladder Carcinoma (TCGA-BLCA)
}
\details{
Obtained via R packages [curatedTCGAData](https://bioconductor.org/packages/release/data/experiment/html/curatedTCGAData.html)
 and [TCGAutils](https://bioconductor.org/packages/release/bioc/html/TCGAutils.html)
}
\keyword{datasets}
